+++ bubo/chat_functions.py
@@ -1,12 +1,18 @@
+import asyncio
 import logging
 import time
+import uuid
 from typing import Optional
 
+import aiohttp
 # noinspection PyPackageRequirements
 from nio import (
-    SendRetryError, RoomInviteError, AsyncClient, ErrorResponse, RoomSendResponse
+    SendRetryError, RoomInviteError, AsyncClient, ErrorResponse, RoomSendResponse,
 )
 from markdown import markdown
+
+from bubo.config import Config
+from bubo.utils import get_request_headers
 
 logger = logging.getLogger(__name__)
 
@@ -97,4 +103,48 @@ async def send_text_to_room(
         else:
             logger.warning(f"Failed to get event_id from send_text_to_room, response: {response}")
     except SendRetryError:
-        logger.exception(f"Unable to send message response to {room_id}")
+        logger.exception(f"Unable to send message to {room_id}")
+
+
+async def send_text_to_room_c2s(
+    config: Config,
+    room_id: str,
+    message: str,
+    notice=True,
+    markdown_convert=True,
+    reply_to_event_id: Optional[str] = None,
+) -> str:
+    """
+    Send text to a matrix room using the C2S API directly.
+    """
+    # Determine whether to ping room members or not
+    msgtype = "m.notice" if notice else "m.text"
+
+    content = {
+        "msgtype": msgtype,
+        "format": "org.matrix.custom.html",
+        "body": message,
+    }
+
+    if markdown_convert:
+        content["formatted_body"] = markdown(message)
+
+    if reply_to_event_id:
+        content["m.relates_to"] = {"m.in_reply_to": {"event_id": reply_to_event_id}}
+
+    try:
+        async with aiohttp.ClientSession() as session:
+            async with session.put(
+                f"{config.homeserver_url}/_matrix/client/r0/rooms/{room_id}/send/m.room.message/{str(uuid.uuid4())}",
+                json=content,
+                headers=get_request_headers(config),
+            ) as response:
+                if response.status == 429:
+                    await asyncio.sleep(3)
+                    return await send_text_to_room_c2s(
+                        config, room_id, message, notice, markdown_convert, reply_to_event_id,
+                    )
+                response.raise_for_status()
+                return (await response.json())["event_id"]
+    except Exception as ex:
+        logger.exception(f"Unable to send C2S message to {room_id}: {ex}")