+++ bubo/utils.py
@@ -10,6 +10,12 @@ from bubo.config import Config
 logger = logging.getLogger(__name__)
 
 
+def get_request_headers(config):
+    return {
+        "Authorization": f"Bearer {config.user_token}",
+    }
+
+
 async def get_users_for_access(client: AsyncClient, config: Config, access_type: str) -> Set:
     if access_type == "admins":
         existing_list = list(set(config.admins[:]))
@@ -31,6 +37,8 @@ async def get_users_for_access(client: AsyncClient, config: Config, access_type:
     return set(users)
 
 
+# TODO remove usage of this wrapper for any matrix-nio calls
+# Reading that code it seems it already handles rate limits 😅
 async def with_ratelimit(client: AsyncClient, method: str, *args, **kwargs):
     func = getattr(client, method)
     response = await func(*args, **kwargs)